Revert "Revert "Revert "Revert "Merge branch 'feature/JNG-2630_Col_value_calculation'""""
This reverts commit 28242a21afde257d15c7cdd1ff84ba58014f8ab3.

+++ lib/src/judo_title.dart
@@ -12,7 +12,7 @@ class JudoTitle extends StatelessWidget {
   }) : super(key: key);
 
   final String text;
-  final int col;
+  final double col;
   final double row;
   final bool stretch;
   final Alignment alignment;
